Distinguish Perl from Prolog, poorly
Perl and Prolog both use the .pl file extension.  Atom seems to prefer
choosing Perl.  However, if the first line contains :- or --> we can be
relatively confident that this is a Prolog file.

Without an ugly hack, Atom only lets us consider the first line of a
file's contents.  In many Prolog files, the first line is a comment
which doesn't contain the distinguishing tokens.  This change doesn't
affect those files.

Ideally, we'd modify Atom's core "first-mate" package teaching its
Grammar::getScore method how to use a "contentMatch" regex (optionally
defined by each grammar) to distinguish Perl from Prolog.

See
https://github.com/atom/first-mate/blob/545c9f56ff1d45f9d1593565ccd32b8ff5bea3c6/src/grammar.coffee#L173
for the area of code that needs to be modified.